Use common status, To #30961382
