+++ api/v1alpha1/status.go
@@ -15,7 +15,11 @@ limitations under the License.
 
 package v1alpha1
 
-import "github.com/fluid-cloudnative/fluid/pkg/common"
+import (
+	"github.com/fluid-cloudnative/fluid/pkg/common"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
 
 // RuntimeStatus defines the observed state of Runtime
 type RuntimeStatus struct {
@@ -107,4 +111,62 @@ type RuntimeStatus struct {
 
 	// CacheStatus represents the total resources of the dataset.
 	CacheStates common.CacheStateList `json:"cacheStates,omitempty"`
+}
+
+type RuntimePhase string
+
+const (
+	RuntimePhaseNone         RuntimePhase = ""
+	RuntimePhaseNotReady     RuntimePhase = "NotReady"
+	RuntimePhasePartialReady RuntimePhase = "PartialReady"
+	RuntimePhaseReady        RuntimePhase = "Ready"
+)
+
+// RuntimeConditionType indicates valid conditions type of a runtime
+type RuntimeConditionType string
+
+// These are valid conditions of a runtime.
+const (
+	// MasterInitialized means the master of runtime is initialized
+	RuntimeMasterInitialized RuntimeConditionType = "MasterInitialized"
+	// MasterReady means the master of runtime is ready
+	RuntimeMasterReady RuntimeConditionType = "MasterReady"
+	// WorkersInitialized means the Workers of runtime is initialized
+	RuntimeWorkersInitialized RuntimeConditionType = "WorkersInitialized"
+	// WorkersReady means the Workers of runtime is ready
+	RuntimeWorkersReady RuntimeConditionType = "WorkersReady"
+	// FusesInitialized means the fuses of runtime is initialized
+	RuntimeFusesInitialized RuntimeConditionType = "FusesInitialized"
+	// FusesReady means the fuses of runtime is ready
+	RuntimeFusesReady RuntimeConditionType = "FusesReady"
+)
+
+const (
+	RuntimeMasterInitializedReason = "Master is initialized"
+	// MasterReady means the master of runtime is ready
+	RuntimeMasterReadyReason = "Master is ready"
+	// WorkersInitialized means the Workers of runtime is initialized
+	RuntimeWorkersInitializedReason = "Workers are initialized"
+	// WorkersReady means the Workers of runtime is ready
+	RuntimeWorkersReadyReason = "Workers are ready"
+	// WorkersInitialized means the Workers of runtime is initialized
+	RuntimeFusesInitializedReason = "Fuses are initialized"
+	// WorkersReady means the Workers of runtime is ready
+	RuntimeFusesReadyReason = "Fuses are ready"
+)
+
+// Condition describes the state of the cache at a certain point.
+type RuntimeCondition struct {
+	// Type of cache condition.
+	Type RuntimeConditionType `json:"type"`
+	// Status of the condition, one of True, False, Unknown.
+	Status corev1.ConditionStatus `json:"status"`
+	// The reason for the condition's last transition.
+	Reason string `json:"reason,omitempty"`
+	// A human readable message indicating details about the transition.
+	Message string `json:"message,omitempty"`
+	// The last time this condition was updated.
+	LastProbeTime metav1.Time `json:"lastProbeTime,omitempty"`
+	// Last time the condition transitioned from one status to another.
+	LastTransitionTime metav1.Time `json:"lastTransitionTime,omitempty"`
 }
